updates to server directory - fixed bcrpyt - reorganized the functions in db/index.js to their own files
